refactor(tui): extract BLOCKS EDIT completion to submodule
